read/elf: split into submodules
